Remove GCC compile warnings
git-svn-id: file:///dev/shm/opensim/Trunk@7561 fc29be9c-2cb0-420d-b520-73dbd18add6e

+++ OpenSim/Actuators/Millard2012EquilibriumMuscle.cpp
@@ -781,7 +781,7 @@ void Millard2012EquilibriumMuscle::
 
         int flag_status    = (int)soln[0];
         double solnErr        = soln[1];
-        double iterations     = (int)soln[2];
+        int iterations     = (int)soln[2];
         double fiberLength    = soln[3];
         double passiveForce   = soln[4];
         double tendonForce    = soln[5];
